pyproject.toml
[project]
name = "hollingsbot3"
version = "0.1.0"
description = "Discord bot with AI integrations, image generation, and more"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "discord.py>=2.5.2",
    "python-dotenv",
    "replicate",
    "aiohttp",
    "celery",
    "redis",
    "pillow",
    "transformers",
    "torch",
    "watchfiles",
    "anthropic",
    "openai",
    "cairosvg",
    "pytesseract",
    "beautifulsoup4",
    "emoji",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "pre-commit>=3.0",
    "mypy>=1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

# =============================================================================
# Ruff - Fast Python linter and formatter (replaces black, isort, flake8, etc.)
# =============================================================================
[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests", "scripts"]

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    "*.pyc",
    "generated",
    "prototype_summarization",
]

[tool.ruff.lint]
# Enable rules
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]

# Ignore rules that conflict with Discord bot patterns or are too noisy
ignore = [
    "E501",   # line too long (handled by formatter, but allow exceptions)
    "B008",   # function call in default argument (common in discord.py)
    "B904",   # raise from in except (often unnecessary)
    "SIM108", # ternary operator (readability preference)
    "UP007",  # Use X | Y for type unions (needs py310+)
    "RUF001", # ambiguous unicode (allow emojis in strings)
    "RUF002", # ambiguous unicode in docstrings
    "RUF003", # ambiguous unicode in comments
]

# Allow autofix for all rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Tests can have more relaxed rules
"tests/**/*.py" = ["S101", "PLR2004"]
# Scripts may have different patterns
"scripts/**/*.py" = ["T201"]  # Allow print statements

[tool.ruff.lint.isort]
known-first-party = ["hollingsbot"]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# =============================================================================
# Mypy - Static type checking
# =============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "generated/",
    "prototype_summarization/",
    "venv/",
    ".venv/",
]

# Relaxed settings for gradual adoption
check_untyped_defs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false

# =============================================================================
# Pytest - Testing configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =============================================================================
# Coverage - Code coverage settings
# =============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/generated/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# =============================================================================
# Bandit - Security linter
# =============================================================================
[tool.bandit]
exclude_dirs = ["tests", "generated", "prototype_summarization", "venv", ".venv"]
skips = [
    "B101",  # assert_used - fine in production for invariants
    "B311",  # random - we use random for non-security purposes
    "B603",  # subprocess_without_shell_equals_true - needed for some operations
]
